refactor(app): split PineApp.swift into menu commands and notifications
Extract the ~600 lines of `CommandGroup` / `CommandMenu` definitions into
`PineAppMenuCommands.swift` as a `Commands`-conforming `PineAppCommands`
struct, and the `Notification.Name` extension into
`PineAppNotifications.swift`. `PineApp.swift` now keeps only `@main`,
Scene wiring, window-close glue, and `AppDelegate`.

Pure mechanical move — no behavior change. PineApp.swift shrinks from
1281 LOC to 804 LOC; future menu edits will diff against a focused file.

Fixes #756
